fix: shift upload dem func  to project logic file
